refactor: DomainStack을 FrontendCertificateStack으로 변경
